+++ app/src/main/java/de/abacuselectronics/aroiorc/ui/custom/divider/DividerType.kt
@@ -1,0 +1,7 @@
+package de.abacuselectronics.aroiorc.ui.custom.divider
+
+enum class DividerType {
+	INDENT,
+	FULL,
+	NONE
+}
